ui - 2019.10.28 (Mon)
- UI building all success

+++ lib/model/response/comments_response.dart
@@ -27,7 +27,7 @@ class CommentsResponse {
 }
 
 class Comment {
-  double timestamp;
+  int timestamp;
   String id;
   String movieId;
   double rating;
